Add react-bootstrap, move to scss
