+++ components/StringToDivs.tsx
@@ -3,6 +3,7 @@ import { FC } from 'react';
 
 /**
  * Placeholder component to convert an array of strings into divs
+ * @deprecated
  * @param param0 
  */
 const StringToDivs: FC<{ data: string[] }> = function StringToDivs({ data }) {
