+++ package.json
@@ -10,9 +10,11 @@
   "dependencies": {
     "@material-ui/core": "^4.11.2",
     "@material-ui/icons": "^4.11.2",
+    "bootstrap": "^4.6.0",
     "gray-matter": "^4.0.2",
     "next": "^10.0.0",
     "react": "17.0.1",
+    "react-bootstrap": "^1.5.1",
     "react-dom": "17.0.1",
     "sass": "^1.32.8"
   },
