+++ pages/_app.tsx
@@ -1,5 +1,6 @@
 import type { AppProps /*, AppContext */ } from 'next/app';
 import Header from '../components/Header';
+import "bootstrap/dist/css/bootstrap.min.css";
 import "../styles/global.scss";
 function MyApp({ Component, pageProps }: AppProps) {
     return (
