+++ styles/colors.scss
@@ -1,0 +1,4 @@
+// Primary root for all styling decisions
+$primary: #c1c1b1;
+$back: #171710;
+$accent: #bc7100;
