+++ styles/global.scss
@@ -1,13 +1,10 @@
-$primary: #c1c1b1;
-$back: #171710;
-$accent: #bc7100;
-
+@use './colors.scss' as c;
 body {
-    background-color: $back;
-    color: $primary;
+    background-color: c.$back;
+    color: c.$primary;
 }
 
 a {
     text-decoration: none;
-    color: $accent;
+    color: c.$accent;
 }
