+++ styles/index.module.scss
@@ -1,4 +1,4 @@
-@use "./global.scss";
+@use "./colors.scss" as c;
 
 .para {
     font-family: sans-serif;
